test(skills): the fuzz ratchet could disarm itself, and its residue note was wrong
Two findings from an adversarial audit of this PR's own claims, both
reproduced before being fixed.

FUZZ_BASELINE is the ratchet's expected value AND the seed list both fuzz
tests are parametrized over, so deleting a key deletes the measurement rather
than failing it. Measured on a scratch copy: setting it to `{}` disarms BOTH
fuzz ratchets and the suite still reports all green, and dropping to a single
seed does the same for the other two. `test_the_fuzz_seed_set_is_pinned` is
deliberately NOT parametrized, so it runs even when the dict is empty.
Mutation verified: emptying the dict now fails a test in each suite instead of
passing silently. Same defect and same remedy as the corpus-size pin.

The residue description was wrong, including the replacement I wrote earlier
in this PR under a sentence beginning "Measured". Re-measured per document:

  seed 1729: five columns of marker padding, an empty item, a thematic break,
             a tab; and it is the ONLY one `--write` still mutates
  seed 4242: maximum two columns of padding, NO empty item, NO thematic break,
             a tab, U+3000 in a fence info string and U+00A0 as a marker
             remainder

So "both carry five-or-more columns of padding" was false, and the empty item
and the thematic break are in the same document rather than split across the
two. The two features seed 4242 actually has appeared in neither description.
The shape now lives in one place, the module docstring, and the fuzz comment
says why it is not repeated: a claim asserted as measured and never
re-measured is worse than none.

Splitting `test_prose_lint.py` because the seed pin took it to 512 lines and
the taste ratchet caught it at 576 against a baseline of 575. Raising a count
baseline to clear a blocked push is forbidden, so the file splits instead, the
way the fence suite already splits its own: the citation replay and the two
CommonMark classes answer to external contracts rather than to this scanner's
detector behaviour. 343 and 199 lines, ratchet back at 575.

464 tests.

Refs #5322
